Fix users, authentication, chart, login

+++ IDS/Controllers/ClientsController.cs
@@ -1,12 +1,15 @@
 ﻿using IDS.Models.Dtos;
 using IDS.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace IDS.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [Authorize]
     public class ClientsController : ControllerBase
     {
         private readonly IClientService _service;
@@ -15,17 +18,28 @@ namespace IDS.Controllers
         {
             _service = clientService;
         }
+        private bool CanManageClients()
+        {
+            var role = User.FindFirst(ClaimTypes.Role)?.Value;
+            var position = User.FindFirst("Position")?.Value;
+            return role == "Admin" || position == "Manager" || position == "CEO" || position == "Project Manager";
+        }
 
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
-            var clients = await _service.GetAllAsync();
+            var position = User.FindFirst("Position")?.Value;
+            var teamIdClaim = User.FindFirst("TeamId")?.Value;
+            int? teamId = teamIdClaim != null ? int.Parse(teamIdClaim) : null;
+
+            var clients = await _service.GetClientsByTeamAsync(position, teamId);
             return Ok(clients);
         }
 
         [HttpGet("{id}")]
         public async Task<IActionResult> GetById(int id)
         {
+
             var client = await _service.GetByIdAsync(id);
             if (client == null)
             {
@@ -46,7 +60,10 @@ namespace IDS.Controllers
         [HttpPost]
         public async Task<IActionResult> Create(CreateClientDto dto)
         {
-            var newId= await _service.CreateAsync(dto);
+            if (!CanManageClients())
+                return Forbid();
+
+            var newId = await _service.CreateAsync(dto);
             return Created();
         }
 
@@ -55,6 +72,9 @@ namespace IDS.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update(int id, CreateClientDto dto)
         {
+            if (!CanManageClients())
+                return Forbid();
+
             var success = await _service.UpdateAsync(id, dto);
             if (!success)
                 return NotFound();
@@ -66,6 +86,9 @@ namespace IDS.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!CanManageClients())
+                return Forbid();
+
             try
             {
                 var deleted = await _service.DeleteAsync(id);
